+++ 42_libft/srcs/libft/ft_numlen_base.c
@@ -6,13 +6,13 @@
 /*   By: lbrangie <lbrangie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/14 11:10:16 by lbrangie          #+#    #+#             */
-/*   Updated: 2018/02/14 11:27:40 by lbrangie         ###   ########.fr       */
+/*   Updated: 2018/03/19 14:51:10 by lbrangie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t			ft_numlen_base(int n, int base)
+size_t			ft_numlen_base(long long n, int base)
 {
 	size_t			len;
